Crate convdate

convdate is a set of tools for converting datetime UTC <=> TAI <=> TT.

If you want to know about executables, look for below documents of binary crates:

tai2utc

tt2utc

utc2tai

utc2tt

This crate provide some features to above binary crates.

Caution

This library crate is being adjusted. There are plans to make disruptive changes in future updates.

Modules

convtbl

Provide tables for time system conversion.

error

Functions

tai2tt

Convert datetime from TAI to TT.

tai2tt_dt

Convert datetime from TAI to TT.

tai2utc

Convert datetime from TAI to UTC.

tai2utc_dt

Convert datetime from TAI to UTC.

tt2tai

Convert datetime from TT to TAI.

tt2tai_dt

Convert datetime from TT to TAI.

tt2utc

Convert datetime from TT to UTC.

tt2utc_dt

Convert datetime from TT to UTC.

utc2tai

Convert datetime from UTC to TAI.

utc2tai_dt

Convert datetime from UTC to TAI.

utc2tt

Convert datetime from UTC to TT.

utc2tt_dt

Convert datetime from UTC to TT.