Function convdate::tai2utc

pub fn tai2utc(
    datetime: &str, 
    utc_tai_table: &UtcTaiTable, 
    dt_fmt: &str
) -> Result<String, Error>

Convert datetime from TAI to UTC.

This function takes leap seconds into account along the argument tai_utc_table.

Arguments

datetime - Datetime in TAI.

utc_tai_table - The conversion table of UTC - TAI

dt_fmt - format of datetime

Returns

Returns the datetime in UTC.

Returns Error if it fail to convert.

Examples

use convdate;
use convdate::convtbl::TaiUtcTable;

// Usually, lines read from the file are used as the argument of `from_lines`.
let tai_utc_table = TaiUtcTable::from_lines(vec!["2017-01-01T00:00:00 37"], "%Y-%m-%dT%H:%M:%S").unwrap();
let utc_tai_table = From::from(&tai_utc_table);

let tai = convdate::tai2utc(
    "2017-01-01T12:00:37.000",
    &utc_tai_table,
    "%Y-%m-%dT%H:%M:%S%.3f");

assert_eq!(tai, Ok("2017-01-01T12:00:00.000".to_string()));

See also

tai2utc_dt - It is same as tai2utc, except that the argument and the result are NaiveDateTime.

tai2utc (Binary crate) - The executable program which do same conversion.